Broke out roguelike.py into modules, moved gear_example into first system test
